+++ src/style/style.ts
@@ -107,6 +107,10 @@ export type StyleOptions = {
      * Forces a full update.
      */
     localIdeographFontFamily?: string | false;
+    /**
+     * Defines an initial global state for the map style. It overrides the defaults defined in the map style, as if {@link Map.setGlobalStateProperty} was called after loading the style.
+     */
+    globalState?: Record<string, any>;
 };
 
 /**
@@ -117,6 +121,10 @@ export type StyleSetterOptions = {
      * Whether to check if the filter conforms to the MapLibre Style Specification. Disabling validation is a performance optimization that should only be used if you have previously validated the values you will be passing to this function.
      */
     validate?: boolean;
+    /**
+     * Defines an initial global state for the map style. It overrides the defaults defined in the map style, as if {@link Map.setGlobalStateProperty} was called after loading the style.
+     */
+    globalState?: Record<string, any>;
 };
 
 /**
@@ -233,6 +241,7 @@ export class Style extends Evented<MapEventType> {
     _layerOrderChanged: boolean;
     _symbolPlacementTriggered: boolean;
     _placedProjectionTransition: number;
+    _initialGlobalState: Record<string, any>;
     _globalState: Record<string, any>;
     crossTileSymbolIndex: CrossTileSymbolIndex;
     pauseablePlacement: PauseablePlacement;
@@ -263,6 +272,7 @@ export class Style extends Evented<MapEventType> {
         this.crossTileSymbolIndex = new CrossTileSymbolIndex();
 
         this._setInitialValues();
+        this._initialGlobalState = options.globalState ?? {};
 
         this._resetUpdates();
 
@@ -299,6 +309,7 @@ export class Style extends Evented<MapEventType> {
         this.tileManagers = {};
         this.zoomHistory = new ZoomHistory();
         this._imagesListDirty = false;
+        this._initialGlobalState = {};
         this._globalState = {};
         this._serializedLayers = {};
         this.stylesheet = null;
@@ -355,8 +366,10 @@ export class Style extends Evented<MapEventType> {
         this._checkLoaded();
 
         const values: Record<string, any> = {};
-        for (const propertyName in newStylesheetState) {
-            values[propertyName] = newStylesheetState[propertyName].default;
+        const propertyNames = new Set([...Object.keys(newStylesheetState ?? {}), ...Object.keys(this._initialGlobalState)]);
+        for (const propertyName of propertyNames) {
+            // Initial global state has priority over global state defaults defined in the map style
+            values[propertyName] = this._initialGlobalState[propertyName] ?? newStylesheetState[propertyName]?.default ?? null;
         }
 
         this._setGlobalStateValues(values);
@@ -875,6 +888,7 @@ export class Style extends Evented<MapEventType> {
             return false;
         }
 
+        this._initialGlobalState = options.globalState ?? {};
         for (const styleChangeOperation of operations.operations) {
             styleChangeOperation();
         }